Move module view extensions to separate files
