+++ Passepartout/Library/Sources/AppUI/Views/Modules/OpenVPNView.swift
@@ -26,24 +26,6 @@
 import PassepartoutKit
 import SwiftUI
 
-extension OpenVPNModule.Builder: ModuleViewProviding {
-    func moduleView(with editor: ProfileEditor) -> some View {
-        OpenVPNView(editor: editor, module: self)
-    }
-}
-
-extension OpenVPNModule.Builder: InteractiveViewProviding {
-    func interactiveView(with editor: ProfileEditor) -> some View {
-        let draft = editor.binding(forModule: self)
-
-        return OpenVPNView.CredentialsView(
-            isInteractive: draft.isInteractive,
-            credentials: draft.credentials,
-            isAuthenticating: true
-        )
-    }
-}
-
 struct OpenVPNView: View {
 
     @ObservedObject
